board: adi: 594: Adjust SPI bootROM baud to 41.6MHz
With the SC594 SOM attached to the ezkit carrier board, the default
SPI2 boot baud for bcode 2 seems to be too fast, generating errors.
Bcode 2 works when the carrier board is disconnected.

The default is 125MHz / (1 + 1) = 62.5MHz. Bcode 0 or 1 is 7.8MHz,
which is painfully slow. This change uses bcode 2 as the base
configuration, but then adjusts the clock to 125MHz / (2 + 1) = 41.6MHz.