Fix `saveenv` in U-Boot, on SC573.
`saveenv` was writing to the wrong place in SPI, so was overwriting
U-Boot itself and rendering the system unbootable, requiring a
reflash from the debug environment.

The error was simply that the offset for the env specified in the
defconfig was overwritten in `sc_adi_common.h`. Removed the lines
that overwrote the relevant config options and it now works fine.